Exported keep_args_alive to dpctl4pybind11 and used that in tensor_py and onemkl_gemv example
